fix(transactions): widen the pesos-to-centavos multiply before it overflows
redemption_logs.purchase_amount is int4 and holds pesos, and the union
scales it by 100 so every amount downstream is centavos. Any sale over
₱21,474,836 overflows 2^31 on that multiply, and Postgres answers 22003
"integer out of range" rather than a row — so one oversized sale on the
books took the entire page down, list and totals together, since both
queries embed the same union.

Cast to bigint before multiplying rather than after. The other branches
supply integer and NULL at that position, so the union resolves to bigint
and mapTransaction reads it back the way it already reads the other
bigint columns.

Nothing guards the write side: the redeem route takes
z.coerce.number().min(0) with no upper bound, so a mistyped amount still
lands in the column. It no longer breaks the page, it just reads as an
absurd sale.

+++ src/server/transactions.ts
@@ -100,7 +100,11 @@ const UNION_SQL = `
     v.voucher_code AS reference,
     v.display_label AS detail,
     rl.note AS note,
-    CASE WHEN rl.purchase_amount IS NULL THEN NULL ELSE rl.purchase_amount * 100 END AS purchase_centavos,
+    -- Widened before the multiply, not after: the column is int4 and holds
+    -- pesos, so any sale over ₱21,474,836 overflowed 2^31 on the way to
+    -- centavos and took the whole page down with a 22003 rather than showing
+    -- the row. Both queries below embed this union, so the totals died with it.
+    CASE WHEN rl.purchase_amount IS NULL THEN NULL ELSE rl.purchase_amount::bigint * 100 END AS purchase_centavos,
     -- Typed, not bare: Postgres unions these pairwise, and an untyped NULL
     -- here meets an untyped NULL in the next branch, resolves the pair to text,
     -- and then cannot be matched to the third branch's integer.
